Labels updating now
First run this program and see that the board has stuff on it from the
demo mode since it's the same code.  Then uncomment the boardUpdate();
call in the initialize method and it will clean it based on where the
board is at which is brand new because the game hasn't been started yet.